requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-docx==1.1.0
docxcompose==1.3.8
docx2txt==0.8
pydantic==2.6.4
httpx==0.27.0
pytest==8.1.1